Fixed Simulation Switching
Hack temporary fix but works for now only handles 2 values for a ui choice. Also realised newtonian physics are incorrectly shared between objects. very heavy objects shouldn't feel an equal pull to a very small object as the very small object feels towards the large object as this leads to very large objects being pulled very fast to very small objects. Currently the simulation is just essentially assuming they're both of the same mass so a star would be pulled towards a planet as much as the planet is pulled towards the star which doesn't work